Extract custom recordings form component
